Split base test and time series test
