+++ samples/snippets/src/test/java/com/example/bigtable/FiltersTest.java
@@ -23,7 +23,7 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class FiltersTest extends BigtableBaseTest {
+public class FiltersTest extends MobileTimeSeriesBaseTest {
 
   @BeforeClass
   public static void beforeClass() throws IOException {
