+++ samples/snippets/src/test/java/com/example/bigtable/ReadsTest.java
@@ -23,7 +23,7 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class ReadsTest  extends BigtableBaseTest {
+public class ReadsTest  extends MobileTimeSeriesBaseTest {
 
   @BeforeClass
   public static void beforeClass() throws IOException {
